extract request serialization code from query protocol generator to standalone module
